5. Refactor code for improved readability and maintainability
